homevork_3.0

+++ spring-web-homework/api/src/main/java/com/geekbrains/spring/web/api/core/ProductDto.java
@@ -1,9 +1,11 @@
 package com.geekbrains.spring.web.api.core;
 
+import java.math.BigDecimal;
+
 public class ProductDto {
     private Long id;
     private String title;
-    private Integer price;
+    private BigDecimal price;
 
     public Long getId() {
         return id;
@@ -21,18 +23,18 @@ public class ProductDto {
         this.title = title;
     }
 
-    public Integer getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 
     public ProductDto() {
     }
 
-    public ProductDto(Long id, String title, Integer price) {
+    public ProductDto(Long id, String title, BigDecimal price) {
         this.id = id;
         this.title = title;
         this.price = price;
